refactor: factor out requests in local_store
